+++ lib/solidus_sendwithus/engine.rb
@@ -2,7 +2,7 @@ module SpreeSendwithus
   class Engine < Rails::Engine
     require 'spree/core'
     isolate_namespace Spree
-    engine_name 'spree_sendwithus'
+    engine_name 'solidus_sendwithus'
 
     config.autoload_paths += %W(#{config.root}/lib)
