+++ lib/solidus_sendwithus/mailer.rb
