+++ lib/solidus_sendwithus/message.rb
