+++ lib/solidus_sendwithus/rspec_support.rb
