+++ spec/lib/solidus_sendwithus/mailer_spec.rb
